requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.1
django-ckeditor==6.2.0
django-crispy-forms==2.1
django-js-asset==2.2.0
django-phone-field==1.8.1
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2